backend/requirements.txt
# SentiGuard Backend Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9      # Required for file uploads

# ── ML ────────────────────────────────────────────────────
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2

# ── QR Scanning ───────────────────────────────────────────
pyzbar==0.1.9
Pillow==10.3.0

# ── Utilities ─────────────────────────────────────────────
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0                # Optional: for making outbound HTTP requests
